requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
lxml==4.9.3
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.1.0